Add files via upload
The .html file is a media page for the All Black fan site that is supposed to contains photos and videos.

The .js is the JavaScript file for a slideshow that I have added to the page. For some reason, the slide show does not render correctly until the left or right button is clicked. Then all of the pictures collapse into the actual slide show view.