Assignment3_Mosinski
The .html file is the media page that contains pictures and a video.

The .js file contains JavaScript code for a carousel slideshow that I have written for the page. It works best in Google Chrome and for some reason, the slideshow does not render correctly until the user clicks one of the arrow buttons to go to the next picture. Then all of the picture collapse into the slideshow.